+++ docs/release-notes.md
@@ -6,6 +6,7 @@
 
 - Make Fields, FieldToColumns, FieldToColumnNames and ColumnsRemapping appropriate mappings.
   You can't do the dirty trick anymore testing for a dict.
+- Fix contracts; it was always an error to access the data attribute directly (just inherited from UserDict).
 
 ### Fixed